backend/requirements.txt
flask
python-dotenv
psycopg2-binary
requests
atproto
flask-cors
gunicorn
sentence-transformers>=2.3.0
transformers<4.39.0
hdbscan
numpy
scikit-learn
umap-learn
matplotlib
plotly